requirements.txt
# FinRL Trading System Requirements
# Clean, organized dependencies with versions

# Core ML/RL Libraries
stable-baselines3==2.0.0
gymnasium==0.29.1
torch>=2.0.0

# Financial Data & Analysis
yfinance==0.2.28
pandas==2.0.3
numpy>=1.24.0,<2.0.0
stockstats==0.6.2
pandas-ta==0.3.14b0

# Trading APIs
alpaca-py==0.13.3
alpaca-trade-api==3.0.2

# Visualization
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Cloud & Deployment
modal==0.56.4

# Data Processing
scipy==1.11.1
scikit-learn==1.3.0

# Utilities
tqdm==4.65.0
colorama==0.4.6

# Testing (optional, for development)
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality (optional, for development)
black==23.7.0
flake8==6.0.0
mypy==1.4.1